Title: 합과 차의 나눗셈, Time: 142ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240207/합과 차의 나눗셈/divide-of-sum-and-sub.java
@@ -4,9 +4,9 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
 
-        double a = sc.nextInt();
-        double b = sc.nextInt();
+        int a = sc.nextInt();
+        int b = sc.nextInt();
 
-        System.out.printf("%.2f", ((a + b) / (a - b)));
+        System.out.printf("%.2f", (double)(a + b) / (a - b));
     }
 }
